LayoutTests/imported/w3c/web-platform-tests/html/syntax/speculative-parsing/generated/document-write/picture-source-no-img.tentative.sub-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_equals: speculative case incorrectly fetched expected "" but got "param-encodingcheck: %C4%9E\r\nAccept: image/webp,image/avif,image/jxl,image/heic,image/heic-sequence,video/*;q=0.8,imag...

    <!-- speculative case in document.write -->
    <picture><source srcset="/html/syntax/speculative-parsing/resources/stash.py?action=put&amp;uuid=fc1496b6-38d0-4556-a4c4-ae6bf83b7659&amp;encodingcheck=&Gbreve;"></picture>


FAIL Speculative parsing, document.write(): picture-source-no-img Unhandled rejection: assert_equals: speculative case incorrectly fetched expected "" but got "param-encodingcheck: %C4%9E\r\nAccept: image/webp,image/avif,image/jxl,image/heic,image/heic-sequence,video/*;q=0.8,image/png,image/svg+xml,image/*;q=0.8,*/*;q=0.5\r\nReferer: http://web-platform.test:8800/html/syntax/speculative-parsing/generated/document-write/picture-source-no-img.tentative.sub.html"
